Extract UnfocusedPreviewReplyTextField
