--- feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/thread/ReplyState.kt
@@ -1,7 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.thread
-
-internal sealed class ReplyState {
-    data class CanCreate(val onCreateReply: () -> Unit) : ReplyState()
-    data class IsSendingReply(val onCancelSendReply: () -> Unit) : ReplyState()
-    object HasSentReply : ReplyState()
-}
